+++ src/internal/hooks/application-loader.ts
@@ -18,6 +18,15 @@ export function useApplicationLoader({ id, serviceKey }: ApplicationLoaderParams
 
   useEffect(() => {
     if (!id) {
+      console.warn('An application ID was not provided. Using default application instead.');
+      setState({ status: 'success', data: DEFAULT_APPLICATION });
+      return;
+    }
+
+    if (!serviceKey) {
+      console.warn(
+        'An application ID was provided, but a service key is missing. Using default application instead.',
+      );
       setState({ status: 'success', data: DEFAULT_APPLICATION });
       return;
     }
@@ -37,7 +46,7 @@ export function useApplicationLoader({ id, serviceKey }: ApplicationLoaderParams
 
 export interface ApplicationLoaderParams {
   id: string | undefined;
-  serviceKey: string;
+  serviceKey: string | undefined;
 }
 
 export type ApplicationLoaderState = Exclude<RequestState<ApplicationInfo>, { status: 'error' }>;