+++ src/models/index.ts
@@ -1,4 +1,6 @@
 export * from './push/notificare-notification';
 export * from './push/notificare-notification-preview-variant';
 export * from './in-app-messaging/notificare-in-app-message';
+export * from './push-onboarding/notificare-push-onboarding-application-info';
+export * from './push-onboarding/notificare-push-onboarding-permission-status';
 export * from './shared/notificare-notification-preview-locale';